requirements.txt
streamlit
faiss-cpu
sentence-transformers
playwright
beautifulsoup4
requests
numpy